Avoid redundant login flows in CLI
We now request all required scopes when authenticating with the Globus
Compute and Auth APIs via the CLI. This avoids prompting the user to run
multiple login flows when starting a multi-user endpoint.
