+++ changelog.d/20250602_121437_30907815+rjmello_multiple_login_flows_sc_42021.rst
@@ -1,0 +1,4 @@
+Bug Fixes
+^^^^^^^^^
+
+- Starting a multi-user endpoint no longer prompts the user to complete multiple login flows.
